% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elgbd-package.R
\docType{package}
\name{elgbd-package}
\alias{elgbd}
\alias{elgbd-package}
\title{elgbd: Empirical Likelihood for General Block Designs}
\description{
Performs hypothesis testing for general block designs with empirical likelihood. The core computational routines are implemented using the 'Eigen' 'C++' library and 'RcppEigen' interface, with 'OpenMP' for parallel computation. Details of the methods are given in Kim, MacEachern, and Peruggia (2023) \doi{10.1080/10485252.2023.2206919}. This work was supported by the U.S. National Science Foundation under Grants No. SES-1921523 and DMS-2015552.
}
\references{
Kim E, MacEachern SN, Peruggia M (2023).
"Empirical likelihood for the analysis of experimental designs."
\emph{Journal of Nonparametric Statistics}, \strong{35}(4), 709--732.
\doi{10.1080/10485252.2023.2206919}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/markean/elgbd}
  \item Report bugs at \url{https://github.com/markean/elgbd/issues}
}

}
\author{
\strong{Maintainer}: Eunseop Kim \email{markean@pm.me} [copyright holder]

Other contributors:
\itemize{
  \item Steven MacEachern [contributor, thesis advisor]
  \item Mario Peruggia [contributor, thesis advisor]
}

}
\keyword{internal}
